+++ java-client/src/test/java/org/openapitools/client/api/UserManagementApiTest.java
@@ -42,7 +42,8 @@ public class UserManagementApiTest {
     @Test
     public void changeUserEnabledTest() throws ApiException {
         OpenUserDTO openUserDTO = null;
-        api.changeUserEnabled(openUserDTO);
+        String operator = null;
+        api.changeUserEnabled(openUserDTO, operator);
         // TODO: test validations
     }
 
@@ -57,7 +58,8 @@ public class UserManagementApiTest {
     public void createOrUpdateUserTest() throws ApiException {
         OpenUserDTO openUserDTO = null;
         Boolean isCreate = null;
-        api.createOrUpdateUser(openUserDTO, isCreate);
+        String operator = null;
+        api.createOrUpdateUser(openUserDTO, isCreate, operator);
         // TODO: test validations
     }